remote/remote_game/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Invitation

class InvitationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Invitation
        fields = ['sender', 'recipient', 'game_mode']
