remote/requirements.txt
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.0.0
gunicorn==21.2.0
